requirements.txt
aiogram==3.25.0
fastapi==0.129.0
uvicorn[standard]==0.41.0
aiohttp==3.13.3
supabase==2.13.0
python-dotenv==1.0.0
pydantic==2.11.3